pyproject.toml
[project]
name = "xmluvation"
version = "1.3.5"
description = "A XML searcher with XPath expressions, with the ability to export matches to a CSV file."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "auto-py-to-exe>=2.47.0",
    "lxml>=6.0.1",
    "pandas>=2.3.1",
    "pyarrow>=21.0.0",
    "pyside6-essentials>=6.10.0",
    "xlsxwriter>=3.2.5",
]

[tool.pyside6-project]
files = ["src/main.py"]
include = [
    "src/config/",
    "src/controllers/",
    "src/dialogs/",
    "src/gui/",
    "src/modules/",
    "src/resources/"
]
